Refactor welcome/completion stage
