--- src/main/java/net/alphalightning/bedwars/setup/map/stages/gamemap/WelcomeStage.java
@@ -1,21 +1,0 @@
-package net.alphalightning.bedwars.setup.map.stages.gamemap;
-
-import net.alphalightning.bedwars.BedWarsPlugin;
-import net.alphalightning.bedwars.feedback.Feedback;
-import net.alphalightning.bedwars.setup.map.MapSetup;
-import net.alphalightning.bedwars.setup.map.stages.Stage;
-import net.kyori.adventure.text.Component;
-import org.bukkit.entity.Player;
-
-public class WelcomeStage extends Stage {
-
-    public WelcomeStage(BedWarsPlugin plugin, Player player, MapSetup setup) {
-        super(plugin, player, setup);
-    }
-
-    @Override
-    public void run() {
-        player.sendMessage(Component.translatable("mapsetup.start"));
-        Feedback.start(player);
-    }
-}
